+++ src/Stack/Dot.hs
@@ -124,7 +124,7 @@ createDependencyGraph dotOpts = do
   installedMap <- fmap snd . fst4 <$> getInstalled menv
                                                    (GetInstalledOpts False False False)
                                                    sourceMap
-  withLoadPackage menv (\loader -> do
+  withLoadPackage (\loader -> do
     let depLoader =
           createDepLoader sourceMap
                           installedMap